Added convenience encoding methods and relative tests. All the work is carried out by the encoder so we just need to make sure that the calls to that are okay.
Google has additional requirements for data sources, not least wrapping them in an additional JS call. :-/ I do not intend to implement the whole interface to supply HTML and CSV. If this formats are required it is likely to be more sensible to produce them directly and not using this library.